Fix failing unit tests
Issue: When run outside pipenv urequests mock imported json dependency from source, avoiding the mock_json added in 99b5fe94. However, with pipenv activated urequests imported the mocked json, which raises OSError instead of JSONDecodeError. The urequests.Response mock only caught JSONDecodeError, so it failed to catch this and replace it with ValueError, resulting in failing tests. Now catches both and replaces with ValueError. It would be really nice if micropython modules would just use the same exceptions so this wouldn't be necessary.