Bug fix: HTTP requests no longer fail
Issue: Request handler detected HTTP by converting the request to string and checking if it started with "GET". This broke due to changes in the str() function between micropython v1.17 and v1.20. In 1.17 str decoded bytes automatically, but on 1.20 str returns a string representation of the bytes, in this case starting with b"GET". The conditional failed to detect this and processed the request as non-HTTP, leading to an exception.