Bug fix: Increment_rule methods no longer accept NaN as argument, return correct error for invalid arguments
Issue: A single try/except block was used to cast both current_rule and argument to required type (int or float). When written this was intended to throw an error when current_rule was enabled or disabled. However, invalid arguments raised exceptions in the same try/except, leading to incorrect error.

MotionSensor: Since validator accepts float and NaN is a valid float, rule could be changed to NaN resulting in a broken reset timer