Bug fix: No longer incorrectly adds turn_on/turn_off/trigger_sensor options to unsupported devices/sensors if supported devices/sensors exist
Issue: The universalOptions array was not copied, a reference to the same object was used for all devices and sensors. When turn_on/turn_off/trigger_sensor were pushed to the array they appeared in the menus of all instances. This also created duplicates if multiple supported devices/sensors existed.