Bug fix: Handle slider moved to 0 correctly
Detected unset rule by checking if rule was falsey, this incorrectly changed text to "Set rule" when rule was 0 or 0.0". If rule was 0 or 0.0 on load the displayed value would be set to enabled, edit button would be shown, and rule input would be standard rule dropdown (with traceback because actual value was 0).