Remove ApiCard fade in/out transition
Intermittent issue (about 1 in 20 times) where ternary in ApiCard fails to change class despite loading state being updated correctly. This results in a blank page (DOM is there but hidden) and requires page refresh. Not worth debugging, the transition isn't necessary (there is a potentially long load time when going to ApiCard, but not when going back to overview) and creates a visual flash when navigating back anyway.