Bug fix: /delete_config no longer leave broken Node model if Config was uploaded
Issue: Originally only intended /delete_config to be used for deleting non-uploaded configs. However, if user creates a config with the same friendly name as an existing node, and clicks overwrite on the duplicate warning modal, a /delete_config request is sent even if the config was uploaded. This left a broken Node in the database and prevented the overview from loading.