Bug fix: edit link no longer breaks when node name in status object doesn't match database
Issue: If new firmware is flashed to existing node it will return a status object with a blank id string. LayoutEmpty used id attribute to get edit config link. This worked when a new config with no devices or sensors was created and uploaded, but fails in more realistic situations.

Now reads node name from URL (same as 84ea1724)