Bug fix: device/sensor enable method no longer runs twice under certain conditions
Issue: If current_rule is "disabled" when enable is called it will call set_rule to replace it with a usable rule (scheduled or default). If the new rule is "enabled" the apply_new_rule method called the enable method a second time, even though the device/sensor was already enabled. This caused sensors to call refresh_group twice, and caused devices to call send 3 times (if their group state was ON).

The apply_new_rule method now only calls enable if the device/sensor is disabled.