Bug fix: revert e8ac9075, add regression test
Isue: The removed lines appeared to be redundant, and the comments did not clearly explain the issue. This scenario only occurs when the create method is called from a timer callback function which is called by loop while iterating self.queue - since the create method does not update the queue that is already being iterated, the loop does not know when the new timer is due and may sleep until the timer after it is due.

This is a somewhat clumsy solution, it would be better to use asyncio.Lock to prevent the queue from being modified until loop finishes iterating (requires refactor, loop is async but functions that modify queue are sync)