Fix intermittent test failure (attempt 3)
Issue: the timer being created expired after 1 second, so it was possible for it to expire and be removed from the schedule during the 10ms yeild that allows the create task to run (asyncio.sleep_ms is not guaranteed to return immediately after the requested period). This seems to be related to the testing environment, not sure if tests are being run in parallel and creating a delay or if there are other tasks on the event loop from previous tests.

Ran full test suite >250 times after fix without the test failing (previously failed every 10-15 runs on dev machine).