Fix docker image
Pin working django version, last build pulled 5.1.3 which caused error when migrations run on startup (django.db.utils.NotSupportedError: SQLite 3.31 or later is required (found 3.27.2)). This is likely related to another dependency (sqlparse?) since sqlite is not installed on the container OS. Importing the sqlite3 package and running  shows version 2.6.0, so not that either.