Normalize IR code timings to NEC protocol spec
Theoretically this should be more memory efficient (only have to allocate 4 ints, instead of dozens when each timing is unique by a few microseconds). In production node with both TV and AC codes total memory consumption is reduced ~300 bytes, no impact on fragmentation but could reduce on a node with more peripherals.